requirements.txt
alembic==1.17.2
altgraph==0.17.4
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
audioop-lts==0.2.2
audioread==3.0.1
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.3
coverage==7.12.0
cryptography==46.0.3
customtkinter==5.2.2
cycler==0.12.1
Cython==3.1.6
darkdetect==0.8.0
decorator==5.2.1
demucs==4.0.1
diffq==0.2.4
distro==1.9.0
dora_search==0.1.12
ecdsa==0.19.1
einops==0.8.1
et_xmlfile==2.0.0
fastapi==0.121.1
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.2
fonttools==4.60.1
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
julius==0.2.7
jwcrypto==1.5.6
kiwisolver==1.4.9
lameenc==1.8.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
msgpack==1.1.2
networkx==3.5
numba==0.62.1
numpy==2.3.4
omegaconf==2.3.0
openai==2.8.1
openpyxl==3.1.5
openunmix==1.3.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pefile==2023.2.7
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pooch==1.8.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==6.3.0
PyPDF2==3.0.1
PySimpleGUI==5.0.10
pytest==9.0.1
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
reportlab==4.4.4
requests==2.32.5
retrying==1.4.2
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.44
sqlmodel==0.0.27
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.49.3
submitit==1.5.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.0
torchaudio==2.9.0
torchcodec==0.8.0
tqdm==4.67.1
treetable==0.2.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
WTForms==3.2.1
